Replaced X86 math with MINGW math assembly (WIP)

+++ src/Musl/src/math/x86_64/fmod.c
@@ -3,12 +3,12 @@
  * This file is part of the mingw-w64 runtime package.
  * No warranty is given; refer to the file DISCLAIMER.PD within this package.
  */
-long double fmodl (long double x, long double y);
+double fmod (double x, double y);
 
-long double
-fmodl (long double x, long double y)
+double
+fmod (double x, double y)
 {
-  long double res = 0.0L;
+  double res = 0.0;
 
   asm volatile (
        "1:\tfprem\n\t"
